requirements.txt
streamlit>=1.34.0,<2.0.0
google-generativeai>=0.3.0,<0.4.0
python-dotenv>=1.0.0